+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## [Unreleased] - 2025-09-16
+### Added
+- Added `get_keyword_full_top` method to show Google's top-100 search results for analyzed keywords
+- Added `get_keywords_info` method to get keyword overview with volume, CPC, competition level, difficulty, and additional metrics for multiple keywords
+- Added `get_keyword_suggestions` method to show search suggestions for keywords found by full-text search with geographic names information
+- Added `get_keyword_top_urls` method to return website pages that rank for the largest amount of analyzed keyword variations with highest traffic
+
 ## [1.0.3] - 2025-09-09
 ### Fixed
 - Error then LLM's is unable to manage `size` parameters in query.